% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourSeasons.R
\name{AmpPred}
\alias{AmpPred}
\title{Seasonal amplitude and predictability within time-series observations}
\usage{
AmpPred(data, info.periods = 4, forecast.periods = 1, cuttoff = 70,
  amp.probs = c(0.975, 0.025), plot = TRUE)
}
\arguments{
\item{{data}}{data.frame as produced by defineSeasons}

\item{{info.periods}}{the number of periods used as prior inforamtion to generate predictions}

\item{{forecast.period}}{the number of periods to be forecasted within each step}

\item{{cuttoff}}{exlude incomplete periods (in the beginning and the end); the cuttoff defines a percetnage of the complete periods.}

\item{{amp.probs}}{the qunatiles used to define the seasonal amplitude.}

\item{{plot}}{logical, of TRUE a plot will be drawn.}
}
\value{
A list with 2 data frames:
\describe{
  \item{summary}{data frame with amplitude and predictability for each seasonal period}
  \item{output}{data frame with all output variables}
}
}
\description{
Function that helps to define the seasonality (periodicity) of the time-series.
}
\details{
Based on time-series analysis, this function first decomposes the data into the following components; seasonal, trend and remainder.
Next, the periodicity of the time-series will be defined by using a cosine function fitted to the seasonal component (using least-square).
The minima of the cosine curve can then be used to separate the periods (nessesary in other functions of the DOSeasons package).
}
\examples{
data(tempYNP)
sTab <- defineSeasons(tempYNP$Date, tempYNP$Tmin, frequency = "daily")
seas <- AmpPred(sTab)
}
\author{
Simeon Lisovski
}

