% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourSeasons.R
\name{defineSeasons}
\alias{defineSeasons}
\title{Define seasonality of time-series}
\usage{
defineSeasons(tm, y, frequency = "daily", plot = TRUE)
}
\arguments{
\item{{tm}}{the time (date; as a POSIXlt or POSIXct class object) of the observation}

\item{{y}}{the observations}

\item{{frequency}}{the frequency of the observations; so far either 'daily', 'weekly' or 'monthly'.}

\item{{plot}}{logical, if TRUE a plot will be produced.}
}
\value{
A data frame with 7 variables:
\describe{
  \item{date}{teh date of the time-series observations}
  \item{y}{the data}
  \item{period}{the number of the period}
  \item{seasonal}{the seasonal component}
  \item{trend}{the trend component}
  \item{remainder}{the remainder component}
  \item{cos.fit}{the best fit of the cosine function}
}
}
\description{
Function that helps to define the seasonality (periodicity) of the time-series.
}
\details{
Based on time-series analysis, this function first decomposes the data into the following components; seasonal, trend and remainder.
Next, the periodicity of the time-series will be defined by using a cosine function fitted to the seasonal component (using least-square).
The minima of the cosine curve can then be used to separate the periods (nessesary in other functions of the DOSeasons package).
}
\examples{
data(tempYNP)
sTab <- defineSeasons(tempYNP$Date, tempYNP$Tmin, frequency = "daily")
}
\author{
Simeon Lisovski
}

