% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourSeasons.R
\docType{data}
\name{tempYNP}
\alias{tempYNP}
\title{Temperature data from Yosemity National Park}
\format{A data frame with 7234 rows and 5 variables:
\describe{
  \item{Sation}{the official name of the weather station}
  \item{Location}{name of location}
  \item{Date}{date of daily summary}
  \item{Tmax}{maximum temperature}
  \item{Date}{minimum temperature}
}}
\source{
\url{https://www.ncdc.noaa.gov/}
}
\description{
Daily minimum and maximum temperatures.
}
\details{
Data source: NOAA (Weather station: Lake Yellowstone, WY US)
}

